[Design]: PlanView 레이아웃 조정 #107
