+++ ForJaeRin.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,0 +1,32 @@
+{
+  "pins" : [
+    {
+      "identity" : "errorkit",
+      "kind" : "remoteSourceControl",
+      "location" : "https://github.com/jaemyeong/ErrorKit.git",
+      "state" : {
+        "revision" : "8e9f8cfc03dfb10701c440f029e7ced1441c10ba",
+        "version" : "0.1.15"
+      }
+    },
+    {
+      "identity" : "pretendard",
+      "kind" : "remoteSourceControl",
+      "location" : "https://github.com/jaemyeong/Pretendard",
+      "state" : {
+        "revision" : "cf518c7b41e35ac68c7faf514bdf0b70d47d83c4",
+        "version" : "0.2.7"
+      }
+    },
+    {
+      "identity" : "swift-log",
+      "kind" : "remoteSourceControl",
+      "location" : "https://github.com/apple/swift-log.git",
+      "state" : {
+        "revision" : "32e8d724467f8fe623624570367e3d50c5638e46",
+        "version" : "1.5.2"
+      }
+    }
+  ],
+  "version" : 2
+}
